+++ src/components/LeaveQueries/Admin.css
@@ -1,3 +1,4 @@
+/* the admin-specific styles (starts from .admin-container) */
 .admin-container {
   padding: 2%;
   font-family: Arial, sans-serif;
