+++ src/components/LeaveQueries/AlertConfirmModals.js
@@ -1,0 +1,34 @@
+// src/components/LeaveQueries/AlertConfirmModals.js
+import React from "react";
+import Modal from "../Modal/Modal";
+
+export default function AlertConfirmModals({
+  alertModal,
+  closeAlert,
+  confirmModal,
+  closeConfirm,
+}) {
+  return (
+    <>
+      <Modal
+        title={alertModal.title}
+        isVisible={alertModal.isVisible}
+        onClose={closeAlert}
+        buttons={[{ label: "OK", onClick: closeAlert }]}
+      >
+        <p>{alertModal.message}</p>
+      </Modal>
+
+      <Modal
+        isVisible={confirmModal.isVisible}
+        onClose={closeConfirm}
+        buttons={[
+          { label: "Cancel", onClick: closeConfirm },
+          { label: "Confirm", onClick: confirmModal.onConfirm },
+        ]}
+      >
+        <p>{confirmModal.message}</p>
+      </Modal>
+    </>
+  );
+}
