+++ src/components/LeaveQueries/LeaveFormModal.js
@@ -1,27 +1,65 @@
+// src/components/LeaveQueries/LeaveFormModal.js
 import React from "react";
 import { MdOutlineCancel } from "react-icons/md";
-import { parseLocalDate, computeRequestedDays } from "./leaveUtils";
+import { getAdvanceNoticeDays } from "./leaveUtils";
 
-/**
- * LeaveFormModal
- * Props:
- * - isFormVisible
- * - formData
- * - handleInputChange
- * - handleSubmit
- * - handleCloseModal
- * - leaveTypeOptions (array of {type,label})
- * - setFormVisible
- */
-const LeaveFormModal = ({
-  isFormVisible,
+export default function LeaveFormModal({
+  isVisible,
+  onClose,
   formData,
+  setFormData,
   handleInputChange,
   handleSubmit,
-  handleCloseModal,
-  leaveTypeOptions = [],
-}) => {
-  if (!isFormVisible) return null;
+  leaveTypeOptions,
+  editingId,
+  // new props
+  showAlert,
+  activePolicy,
+  defaultLeaveSettings,
+}) {
+  if (!isVisible) return null;
+
+  // Called when user selects a leave type — we first update form state via the provided handler,
+  // then compute the applicable advance-notice days and show an alert immediately if needed.
+  const onLeaveTypeChange = (e) => {
+    // update the form state as before
+    handleInputChange(e);
+
+    // if editing existing request, do not show the notice (skip)
+    if (editingId) return;
+
+    const selectedType = String(e.target.value || "").toLowerCase();
+    if (!selectedType) return;
+
+    // find setting from active policy first (if any)
+    let setting =
+      (activePolicy?.leave_settings || []).find(
+        (s) => String(s.type || "").toLowerCase() === selectedType
+      ) || null;
+
+    // fallback to default settings if not found or no active policy
+    if (!setting && Array.isArray(defaultLeaveSettings)) {
+      setting = defaultLeaveSettings.find(
+        (s) => String(s.type || "").toLowerCase() === selectedType
+      );
+    }
+
+    if (!setting) return;
+
+    const noticeDays = getAdvanceNoticeDays(setting);
+    if (noticeDays > 0) {
+      // Show explicit message depending on whether there's an active policy or not
+      if (!activePolicy) {
+        showAlert(
+          `By default, a ${setting.type} request requires at least ${noticeDays} day(s) advance. Please choose a start date at least ${noticeDays} day(s) after today.`
+        );
+      } else {
+        showAlert(
+          `This "${setting.type}" leave requires at least ${noticeDays} day(s) advance. Please choose a start date at least ${noticeDays} day(s) after today.`
+        );
+      }
+    }
+  };
 
   return (
     <div className="leave-modal">
@@ -29,16 +67,15 @@ const LeaveFormModal = ({
         <form className="leave-form" onSubmit={handleSubmit}>
           <div className="leave-form-header">
             <h2>Leave Request Form</h2>
-            <MdOutlineCancel className="icon" onClick={handleCloseModal} />
+            <MdOutlineCancel className="icon" onClick={onClose} />
           </div>
-
           <div className="leave-form-grid">
             <div className="leave-form-group">
               <label>Type of Leave</label>
               <select
                 name="leavetype"
                 value={formData.leavetype}
-                onChange={handleInputChange}
+                onChange={onLeaveTypeChange}
               >
                 <option value="">Select</option>
                 {leaveTypeOptions.map((opt) => (
@@ -48,7 +85,6 @@ const LeaveFormModal = ({
                 ))}
               </select>
             </div>
-
             <div className="leave-form-group">
               <label>Start Date</label>
               <input
@@ -60,7 +96,6 @@ const LeaveFormModal = ({
                 min={new Date().toISOString().split("T")[0]}
               />
             </div>
-
             <div className="leave-form-group">
               <label>End Date</label>
               <input
@@ -72,7 +107,6 @@ const LeaveFormModal = ({
                 required
               />
             </div>
-
             <div className="leave-form-group">
               <label>Half/Full Day</label>
               <select
@@ -89,7 +123,6 @@ const LeaveFormModal = ({
                 <option value="Half Day">Half Day</option>
               </select>
             </div>
-
             <div className="leave-form-group">
               <label>Leave Reason</label>
               <input
@@ -101,23 +134,16 @@ const LeaveFormModal = ({
               />
             </div>
           </div>
-
           <div className="leave-form-actions">
-            <button
-              type="button"
-              className="leave-cancel"
-              onClick={handleCloseModal}
-            >
+            <button type="button" className="leave-cancel" onClick={onClose}>
               Cancel
             </button>
             <button type="submit" className="leave-save">
-              Submit
+              {editingId ? "Update" : "Submit"}
             </button>
           </div>
         </form>
       </div>
     </div>
   );
-};
-
-export default LeaveFormModal;
+}
