+++ src/components/LeaveQueries/LeaveRequest.css
@@ -1,3 +1,4 @@
+/* entire CSS from start up to (but not including) the .admin-container block */
 .leave-container {
   margin: 1% 3%;
   padding: 0%;
